wifi: iwlwifi: Fix spurious packet drops with RSS
When RSS is used and one of the RX queues lags behind others by more than
2048 frames, then new frames arriving on the lagged RX queue are
incorrectly treated as old rather than new by the reorder buffer, and are
thus spuriously dropped. This is because the reorder buffer treats frames
as old when they have an SN that is more than 2048 away from the head SN,
which causes the reorder buffer to drop frames that are actually valid.

The odds of this occurring naturally increase with the number of
RX queues used, so CPUs with many threads are more susceptible to
encountering spurious packet drops caused by this issue.

As it turns out, the firmware already detects when a frame is either old or
duplicated and exports this information, but it's currently unused. Using
these firmware bits to decide when frames are old or duplicated fixes the
spurious drops.

Johannes mentions that the 9000 series' firmware doesn't support these
bits, so disable RSS on the 9000 series chipsets since they lack a
mechanism to properly detect old and duplicated frames.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=216044
CC: stable@vger.kernel.org # 4.7+
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>